+++ pydoll/browser/__init__.py
@@ -1,4 +1,5 @@
 from pydoll.browser.chrome import Chrome
 from pydoll.browser.edge import Edge
+from pydoll.browser.fingerprint import FINGERPRINT_MANAGER, FingerprintManager
 
-__all__ = ['Chrome', 'Edge']
+__all__ = ['Chrome', 'Edge', 'FINGERPRINT_MANAGER', 'FingerprintManager']